+++ src/Kronika/Time.php
@@ -600,11 +600,7 @@ final readonly class Time implements Unit
      */
     public function compareTo(self|TimeUnit $other, Precision $precision = Precision::Micro): Compared
     {
-        return match ($precision) {
-            Precision::Micro => $this->instant()->compareTo($this->normalize($other)->instant()),
-            Precision::Second => $this->resetMicro()->compareTo($this->normalize($other)->resetMicro()),
-            Precision::Minute => $this->resetSecond()->compareTo($this->normalize($other)->resetSecond()),
-        };
+        return $this->instant()->compareTo($this->normalize($other)->instant(), $precision);
     }
 
     /**